rename class and add explanation to shadow
