+++ app/src/org/commcare/models/database/app/models/UserKeyRecord.java
@@ -5,8 +5,7 @@ package org.commcare.models.database.app.models;
 
 import org.commcare.CommCareApp;
 import org.commcare.models.database.SqlStorage;
-import org.commcare.models.encryption.CryptUtil;
-import org.commcare.models.encryption.StringWrapper;
+import org.commcare.models.encryption.ByteEncrypter;
 import org.commcare.models.framework.Persisted;
 import org.commcare.models.framework.Persisting;
 import org.commcare.models.framework.Table;
@@ -277,7 +276,7 @@ public class UserKeyRecord extends Persisted {
     }
 
     public void assignPinToRecord(String pin, String password) {
-        this.passwordWrappedByPin = StringWrapper.wrapByteArrayWithString(password.getBytes(), pin);
+        this.passwordWrappedByPin = ByteEncrypter.wrapByteArrayWithString(password.getBytes(), pin);
     }
 
     public byte[] getWrappedPassword() {
@@ -298,7 +297,7 @@ public class UserKeyRecord extends Persisted {
      * is not valid to unwrap the wrapped password
      */
     public String getUnhashedPasswordViaPin(String pin) {
-        byte[] unwrapped = StringWrapper.unwrapByteArrayWithString(this.passwordWrappedByPin, pin);
+        byte[] unwrapped = ByteEncrypter.unwrapByteArrayWithString(this.passwordWrappedByPin, pin);
         if (unwrapped == null) {
             // If the pin could not unwrap the password, just return null
             return null;
@@ -320,7 +319,7 @@ public class UserKeyRecord extends Persisted {
 
     public byte[] unWrapKey(String password) {
         if (isPasswordValid(password)) {
-            return StringWrapper.unwrapByteArrayWithString(getEncryptedKey(), password);
+            return ByteEncrypter.unwrapByteArrayWithString(getEncryptedKey(), password);
         } else {
             //throw exception?
             return null;
