+++ app/src/org/commcare/models/database/user/DemoUserBuilder.java
@@ -10,7 +10,7 @@ import org.commcare.models.database.ConcreteAndroidDbHelper;
 import org.commcare.models.database.SqlStorage;
 import org.commcare.models.database.app.models.UserKeyRecord;
 import org.commcare.models.encryption.CryptUtil;
-import org.commcare.models.encryption.StringWrapper;
+import org.commcare.models.encryption.ByteEncrypter;
 import org.javarosa.core.model.User;
 import org.javarosa.core.util.PropertyUtils;
 
@@ -78,7 +78,7 @@ public class DemoUserBuilder {
     }
 
     private UserKeyRecord writeNewKeyRecord() {
-        byte[] encryptedKey = StringWrapper.wrapByteArrayWithString(randomKey, password);
+        byte[] encryptedKey = ByteEncrypter.wrapByteArrayWithString(randomKey, password);
 
         UserKeyRecord keyRecord =
                 new UserKeyRecord(username, passwordHash, encryptedKey,
