+++ app/src/org/commcare/xml/KeyRecordParser.java
@@ -2,8 +2,7 @@ package org.commcare.xml;
 
 import org.commcare.data.xml.TransactionParser;
 import org.commcare.models.database.app.models.UserKeyRecord;
-import org.commcare.models.encryption.CryptUtil;
-import org.commcare.models.encryption.StringWrapper;
+import org.commcare.models.encryption.ByteEncrypter;
 import org.commcare.utils.Base64;
 import org.commcare.utils.Base64DecoderException;
 import org.javarosa.xml.util.InvalidStructureException;
@@ -69,7 +68,7 @@ public abstract class KeyRecordParser extends TransactionParser<ArrayList<UserKe
                 throw new InvalidStructureException("Invalid AES key in key record", parser);
             }
 
-            byte[] wrappedKey = StringWrapper.wrapByteArrayWithString(theKey, currentpwd);
+            byte[] wrappedKey = ByteEncrypter.wrapByteArrayWithString(theKey, currentpwd);
             UserKeyRecord record = new UserKeyRecord(username, UserKeyRecord.generatePwdHash(currentpwd), wrappedKey, valid, expires, uuid, UserKeyRecord.TYPE_NEW);
 
             keyRecords.add(record);
