+++ unit-tests/resources/robolectric.properties
@@ -1,5 +1,5 @@
 manifest=../app/AndroidManifest.xml
 res=../app/res
 assets=../app/assets
-shadows=org.commcare.android.shadows.SQLiteDatabaseNative,org.commcare.android.shadows.CommCareShadowLog,org.commcare.android.shadows.ShadowAsyncTaskNoExecutor,org.commcare.android.shadows.StringWrapperShadow
+shadows=org.commcare.android.shadows.SQLiteDatabaseNative,org.commcare.android.shadows.CommCareShadowLog,org.commcare.android.shadows.ShadowAsyncTaskNoExecutor,org.commcare.android.shadows.ByteEncrypterShadow
 sdk=21
